retorrenter refactor
Initial effort to refactor the retorrenter, to avoid the
one-giant-function pattern.

Rather than the previous model, which relied upon the order of paths
returned from listdir() for separate src and dst lists, build a dict of
src : dst pairs, and use the dict to build the command bundles.

This refactor is likely to break some (many?) of the untested areas of
the codebase, which is to say, maybe a lot of it, since it's an attempt
to unify terminology and variable names, which used to be misleading
(filenames, paths etc.)

Fix OP/ED handling, finish refactor